+++ cmb_protocol/server.py
@@ -9,9 +9,10 @@ from cmb_protocol.coding import Encoder
 from cmb_protocol.connection import ServerSideConnection
 from cmb_protocol.constants import MAXIMUM_TRANSMISSION_UNIT, SYMBOLS_PER_BLOCK, RESOURCE_ID_STRUCT_FORMAT
 from cmb_protocol.packets import PacketType, RequestResource
-from cmb_protocol.helpers import spawn_child_nursery, get_logger, set_listen_address, set_remote_address, get_ip_family
+from cmb_protocol.helpers import spawn_child_nursery, get_ip_family
+from cmb_protocol import log_util
 
-logger = get_logger(__name__)
+logger = log_util.get_logger(__name__)
 
 
 async def run_accept_loop(udp_sock, resource_id, encoders):
@@ -20,8 +21,8 @@ async def run_accept_loop(udp_sock, resource_id, encoders):
         while True:
             try:
                 data, address = await udp_sock.recvfrom(2048)
-                set_remote_address(address)
                 packet = PacketType.parse_packet(data)
+                log_util.set_remote_address(address)
                 logger.debug('Received %s', packet)
             except (ConnectionResetError, ConnectionRefusedError):
                 # ignore error as we can't infer which send operation failed
@@ -36,7 +37,9 @@ async def run_accept_loop(udp_sock, resource_id, encoders):
                     child_nursery, shutdown_trigger = await spawn_child_nursery(nursery, shutdown_timeout=3)
 
                     def shutdown():
+                        # trigger child nursery timeout
                         shutdown_trigger.set()
+                        # remove from dict to prevent handling of future packets
                         del connections[address]
                         logger.debug('Closed connection')
 
@@ -56,7 +59,8 @@ async def serve(addresses, resource_id, encoders):
     async with trio.open_nursery() as nursery:
         for address in addresses:
             async def _serve():
-                set_listen_address(address)
+                log_util.set_listen_address(address)
+
                 with socket.socket(family=get_ip_family(address), type=socket.SOCK_DGRAM) as udp_sock:
                     await udp_sock.bind(address)
                     logger.info('Started listening')
